+++ src/extension.ts
@@ -5,11 +5,11 @@ import JsonPathExtension from './jsonPathExtension';
 
 export function activate(context: vscode.ExtensionContext) {
 
-    const jsonPathPlainText = vscode.commands.registerCommand('extension.jsonpathplaintext', () => {
+    const jsonPathPlainText = vscode.commands.registerCommand('jsonPathExtract.queryToPlainText', () => {
         const jpe = new JsonPathExtension(false);
         jpe.run();
     });
-    const jsonPathJson = vscode.commands.registerCommand('extension.jsonpathjson', () => {
+    const jsonPathJson = vscode.commands.registerCommand('jsonPathExtract.queryToJson', () => {
         const jpe = new JsonPathExtension(true);
         jpe.run();
     });