+++ src/test/extension.test.ts
@@ -1,22 +1,22 @@
-//
-// Note: This example test is leveraging the Mocha test framework.
-// Please refer to their documentation on https://mochajs.org/ for help.
-//
+'use strict';
 
-// The module 'assert' provides assertion methods from node
-import * as assert from 'assert';
+import { assert } from 'chai';
+import * as vscode from 'vscode';
+import * as myExtension from '../extension';
 
-// You can import and use all API from the 'vscode' module
-// as well as import your extension to test it
-// import * as vscode from 'vscode';
-// import * as myExtension from '../extension';
+describe("Extension tests", function() {
+    it("should export activate function", function() {
+        assert.typeOf(myExtension.activate, "function");
+    });
 
-// Defines a Mocha test suite to group tests of similar kind together
-suite("Extension Tests", function () {
+    it("should export deactivate function", function() {
+        assert.typeOf(myExtension.deactivate, "function");
+    });
 
-    // Defines a Mocha unit test
-    test("Something 1", function() {
-        assert.equal(-1, [1, 2, 3].indexOf(5));
-        assert.equal(-1, [1, 2, 3].indexOf(0));
+    it("should activate", async function() {
+        await vscode.commands.executeCommand('jsonPathExtract.queryToPlainText');
+        const list = await vscode.commands.getCommands();
+        assert.include(list, 'jsonPathExtract.queryToPlainText');
+        assert.include(list, 'jsonPathExtract.queryToJson');
     });
 });