+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to the **JSONPath Extract** extension will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Integration tests
 
 ## [0.1.0] - 2018-05-28
 ### Added